+++ empire/server/api/v2/module/module_dto.py
@@ -18,7 +18,8 @@ def domain_to_dto_module(module: EmpireModule, uid: str):
                     "value": x[1].value,
                     "strict": x[1].strict,
                     "suggested_values": x[1].suggested_values,
-                    "value_type": to_value_type(x[1].value),
+                    # todo expand to listener, stager, etc
+                    "value_type": to_value_type(x[1].value, x[1].type),
                 },
             ),
             options.items(),