+++ empire/server/modules/powershell/code_execution/invoke_assembly.py
@@ -3,7 +3,6 @@ from __future__ import print_function
 from builtins import object, str
 from typing import Dict
 
-from empire.server.core.db.base import SessionLocal
 from empire.server.core.module_models import EmpireModule
 from empire.server.utils.module_util import handle_error_message
 
@@ -56,9 +55,7 @@ class Module(object):
             return handle_error_message(err)
 
         try:
-            encode_assembly = main_menu.downloadsv2.get_all(
-                SessionLocal(), None, params["File"]
-            )[0][0].get_base64_file()
+            encode_assembly = params["File"].get_base64_file()
         except Exception:
             return handle_error_message(
                 "[!] Could not read .NET assembly path at: " + str(params["Arguments"])